backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.128.8",
    "finnhub-python>=2.4.27",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.39.0",
    "yfinance>=1.2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
